improve formatting of release script

+++ scripts/release-version.sh
@@ -6,12 +6,24 @@ set -xe
 current_version="$(grep '^version = ' Cargo.toml | head -1 | cut -d '"' -f2)"
 new_version="$1"
 
+if [ -z "$new_version" ]; then
+    echo "New version required as argument"
+    exit 1
+fi
+
+echo ">>> Bumping version"
+
 action_prefix="untitaker\\/hyperlink@"
 sed -i.bak "s/$action_prefix$current_version/$action_prefix$new_version/" README.md
 rm README.md.bak
 sed -i.bak "s/version = \"$current_version\"/version = \"$new_version\"/" Cargo.toml
 rm Cargo.toml.bak
+
+echo ">>> Running tests"
 cargo build
+cargo test
+
+echo ">>> Commit"
 
 git add README.md
 git add Cargo.toml
@@ -23,6 +35,7 @@ git show HEAD
 set +x
 
 echo "things left to do:"
+echo "  cargo publish"
 echo "  git push"
 echo "  git push origin $new_version"
 echo "  uncheck and check 'Publish to Marketplace' property of the new release"
